backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
motor>=3.6,<4
pymongo>=4.10,<5
pydantic-settings>=2.6,<3
PyJWT>=2.10,<3
bcrypt>=4.2,<5
email-validator>=2.2,<3
